+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 
-- overwriteRequestQueryParams - Auto-enable query parameters when overwrite value is set
-
+- overwriteRequestQueryParams - Auto-enable query parameters when overwrite value is set (#559)
+- overwriteRequestHeaders - Auto-enable headers when overwrite value is set (#559)
 
 ## v1.26.2 - (2024-01-28)